Added things that we apparently should have
Apparently, credit card numbers are always between 12 and 15 digits
rather than any length, as I was told, so my thing now checks for that.
Also, a credit card with no apparent type is invalid even though in the
types array the driver says “other,” implying that there are other
kinds of credit cards, so my thing now checks for that, too.